fix: nested routes in ssg
