+++ packages/solutions/app-tools/src/analyze/templates.ts
@@ -267,6 +267,7 @@ export const fileSystemRoutes = async ({
         if (route.isRoot) {
           rootLayoutCode = `import RootLayout from '${route._component}'`;
           component = `RootLayout`;
+          // Todo SSG
         } else if (ssrMode === 'string') {
           lazyImport = `() => import(/* webpackChunkName: "${route.id}" */  '${route._component}')`;
           component = `loadable(${lazyImport})`;
