+++ packages/solutions/app-tools/src/types/config/output.ts
@@ -1,6 +1,7 @@
 import type { BuilderConfig } from '@modern-js/builder-webpack-provider';
 import type { BuilderConfig as RsBuilderConfig } from '@modern-js/builder-rspack-provider';
 import type { SharedOutputConfig as BuilderSharedOutputConfig } from '@modern-js/builder-shared';
+import type { SSGConfig } from '@modern-js/types';
 import { UnwrapBuilderConfig } from '../utils';
 
 export type BuilderOutputConfig = UnwrapBuilderConfig<BuilderConfig, 'output'>;
@@ -22,30 +23,9 @@ export interface RsOutputUserConfig
   extends RsBuilderOutputConfig,
     SharedOutputConfig {}
 
-export type SSGRouteOptions =
-  | string
-  | {
-      url: string;
-      output?: string;
-      params?: Record<string, any>[];
-      headers?: Record<string, any>;
-    };
-
-export type SSGSingleEntryOptions =
-  | boolean
-  | {
-      preventDefault?: string[];
-      headers?: Record<string, any>;
-      routes?: SSGRouteOptions[];
-    };
-
-export type SSGMultiEntryOptions = Record<string, SSGSingleEntryOptions>;
-
-export type SSGConfig =
-  | boolean
-  | SSGSingleEntryOptions
-  | SSGMultiEntryOptions
-  | ((
-      entryName: string,
-      payload: { baseUrl?: string },
-    ) => SSGSingleEntryOptions);
+export type {
+  SSGRouteOptions,
+  SSGSingleEntryOptions,
+  SSGMultiEntryOptions,
+  SSGConfig,
+} from '@modern-js/types';
