+++ packages/toolkit/types/cli/index.d.ts
@@ -88,4 +88,32 @@ export interface HtmlTemplates {
   [name: string]: string;
 }
 
+export type SSGRouteOptions =
+  | string
+  | {
+      url: string;
+      output?: string;
+      params?: Record<string, any>[];
+      headers?: Record<string, any>;
+    };
+
+export type SSGSingleEntryOptions =
+  | boolean
+  | {
+      preventDefault?: string[];
+      headers?: Record<string, any>;
+      routes?: SSGRouteOptions[];
+    };
+
+export type SSGMultiEntryOptions = Record<string, SSGSingleEntryOptions>;
+
+export type SSGConfig =
+  | boolean
+  | SSGSingleEntryOptions
+  | SSGMultiEntryOptions
+  | ((
+      entryName: string,
+      payload: { baseUrl?: string },
+    ) => SSGSingleEntryOptions);
+
 export type { Merge } from 'type-fest';
