+++ packages/toolkit/utils/src/is/index.ts
@@ -1,7 +1,12 @@
 import fs from 'fs';
 import path from 'path';
+import type { SSGMultiEntryOptions } from '@modern-js/types';
 import { createDebugger } from '../debug';
 import { isDev } from './nodeEnv';
+
+interface EntryPoint {
+  entryName: string;
+}
 
 const debug = createDebugger('judge-depExists');
 /**
@@ -101,6 +106,25 @@ export const isRouterV5 = (config: {
 }) =>
   typeof config.runtime?.router !== 'boolean' &&
   config?.runtime?.router?.mode === 'react-router-5';
+
+export const isSSGEntry = (
+  config: any,
+  entryName: string,
+  entrypoints: EntryPoint[],
+) => {
+  const ssgConfig = config.output.ssg;
+  const useSSG = isSingleEntry(entrypoints)
+    ? Boolean(ssgConfig)
+    : ssgConfig === true ||
+      typeof (ssgConfig as Array<unknown>)?.[0] === 'function' ||
+      Boolean((ssgConfig as SSGMultiEntryOptions)?.[entryName]);
+
+  return useSSG;
+};
+
+// TODO: remove hard code 'main'
+export const isSingleEntry = (entrypoints: EntryPoint[]) =>
+  entrypoints.length === 1 && entrypoints[0].entryName === 'main';
 
 export * from './nodeEnv';
 export * from './platform';
